# Add zlib-ng and refactor gzip open functions.