I added zlib-ng. This is great, but now we also have three competing  library implementations for opening deflate compressed files:
- zlib
- ISA-L
- zlib-ng

Each of these has their own implementation programs as well. This means the choice for the "best" application becomes more situational. Also it is hard to correctly set all the initialization options.

I have therefore refactored the code. A list of preferred implementations is made. Then using a for loop and a try except the list is tried in order for an application that works. Application-specific options are set using functool.partial

This leads to more verbose code, but also in much simpler logic. Therefore this solution scales better. Since I expect crabz to be added soon, it is much easier to do that in this codebase.

Also this will work wonders for coverage as all the individual PipedCompression Readers and Writers are tested separately anyway and (almost) all the code in the opening functions is touched as there is only a single opening code line within a for loop.
It should be much more robust this way.