@rhpvorderman 2023-02-03
benchmarks again:
Before 
```
Compressed at level 1 with 0 threads; filesize: 405313510, time: 3.0943868160247803 seconds
Compressed at level 2 with 0 threads; filesize: 404989250, time: 3.2503600120544434 seconds
Compressed at level 3 with 0 threads; filesize: 407253483, time: 8.559628248214722 seconds
Compressed at level 4 with 0 threads; filesize: 386930500, time: 22.203505277633667 seconds
Compressed at level 5 with 0 threads; filesize: 374207207, time: 39.5833375453949 seconds
Compressed at level 6 with 0 threads; filesize: 355347468, time: 98.56518363952637 seconds
Compressed at level 1 with 1 threads; filesize: 405313274, time: 3.134148597717285 seconds
Compressed at level 2 with 1 threads; filesize: 404989061, time: 3.2902400493621826 seconds
Compressed at level 3 with 1 threads; filesize: 407251397, time: 8.680283308029175 seconds
Compressed at level 4 with 1 threads; filesize: 387016236, time: 22.396305322647095 seconds
Compressed at level 5 with 1 threads; filesize: 374290891, time: 40.33376383781433 seconds
Compressed at level 6 with 1 threads; filesize: 355500797, time: 100.41857743263245 seconds
Compressed at level 1 with 4 threads; filesize: 405313274, time: 3.1683871746063232 seconds
Compressed at level 2 with 4 threads; filesize: 404989061, time: 3.349888563156128 seconds
Compressed at level 3 with 4 threads; filesize: 407251397, time: 8.81507396697998 seconds
Compressed at level 4 with 4 threads; filesize: 387016236, time: 5.959070682525635 seconds
Compressed at level 5 with 4 threads; filesize: 374290891, time: 11.105888366699219 seconds
Compressed at level 6 with 4 threads; filesize: 355500797, time: 27.333863973617554 seconds
```
After:
```
Compressed at level 1 with 0 threads; filesize: 405313510, time: 3.118347406387329 seconds
Compressed at level 2 with 0 threads; filesize: 404989250, time: 3.301008939743042 seconds
Compressed at level 3 with 0 threads; filesize: 376289859, time: 20.477195024490356 seconds
Compressed at level 4 with 0 threads; filesize: 368802018, time: 17.948566198349 seconds
Compressed at level 5 with 0 threads; filesize: 359170597, time: 24.44700598716736 seconds
Compressed at level 6 with 0 threads; filesize: 347379910, time: 41.21350431442261 seconds
Compressed at level 1 with 1 threads; filesize: 405313274, time: 3.2212672233581543 seconds
Compressed at level 2 with 1 threads; filesize: 404989061, time: 3.3104476928710938 seconds
Compressed at level 3 with 1 threads; filesize: 376289859, time: 20.57276463508606 seconds
Compressed at level 4 with 1 threads; filesize: 368802018, time: 17.815391063690186 seconds
Compressed at level 5 with 1 threads; filesize: 359170597, time: 24.02311873435974 seconds
Compressed at level 6 with 1 threads; filesize: 347379910, time: 40.91991400718689 seconds
Compressed at level 1 with 4 threads; filesize: 405313274, time: 3.207160472869873 seconds
Compressed at level 2 with 4 threads; filesize: 404989061, time: 3.3800032138824463 seconds
Compressed at level 3 with 4 threads; filesize: 376289859, time: 20.435675144195557 seconds
Compressed at level 4 with 4 threads; filesize: 368802018, time: 17.669416427612305 seconds
Compressed at level 5 with 4 threads; filesize: 359170597, time: 23.864430904388428 seconds
Compressed at level 6 with 4 threads; filesize: 347379910, time: 40.691060066223145 seconds
```
I am much more happy with the compression gradient. Level 3 is now actually an improvement over level 2 (albeit much slower). Zlib-ng does some really nice work. Only threads are missing, but it looks crabz will provide these when ready.

I am also quite happy about how this PR turns out technically. It should be quite easy to add crabz after this and the logic has gotten a bit simpler.